deps: add pkg-config to coarray with CMake fallback
make coarray in its own file for clarity
